+++ devenv-nix-backend/devenv.nix
@@ -4,15 +4,20 @@
 , config
 , ...
 }:
+let
+  patchedNix = import ../nix/patched-nix.nix {
+    nix = inputs.nix.packages.${pkgs.stdenv.system}.nix;
+  };
+in
 {
   packages = [
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-expr-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-store-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-util-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-flake-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-cmd-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-fetchers-c
-    inputs.nix.packages.${pkgs.stdenv.system}.nix-main-c
+    patchedNix.libs.nix-expr-c
+    patchedNix.libs.nix-store-c
+    patchedNix.libs.nix-util-c
+    patchedNix.libs.nix-flake-c
+    patchedNix.libs.nix-cmd-c
+    patchedNix.libs.nix-fetchers-c
+    patchedNix.libs.nix-main-c
     pkgs.boehmgc
     pkgs.rustPlatform.bindgenHook
   ];