+++ flake.nix
@@ -85,13 +85,16 @@
       packages = forAllSystems (
         system:
         let
+          patchedNix = import ./nix/patched-nix.nix {
+            nix = inputs.nix.packages.${system}.nix;
+          };
           overlays = [
             inputs.rust-overlay.overlays.default
             (final: prev: {
               inherit (inputs.cachix.packages.${system}) cachix;
               # Use nix-cli to skip building nix-manual, but keep libs for C bindings
               nix = inputs.nix.packages.${system}.nix-cli // {
-                inherit (inputs.nix.packages.${system}.nix) libs;
+                inherit (patchedNix) libs;
               };
               nixd = inputs.nixd.packages.${system}.nixd.override { llvmStatic = true; };
               crate2nix = final.callPackage "${inputs.crate2nix}/crate2nix/default.nix" { };