+++ nix/patched-nix.nix
@@ -1,0 +1,5 @@
+{ nix }:
+
+nix.appendPatches [
+  ./nix-build-environment-ca-derivations.patch
+]